Genericise `DocumentView` and make `DocumentHistoryView` use that class
Introduce a new `DocumentView` class inheriting from `TemplateView`,
which handles all the standard loading of context used in EUI pages
which deal with `Document`s.
