+++ judgments/utils/view_helpers.py
@@ -1,5 +1,6 @@
 from typing import Any
 
+import ds_caselaw_utils as caselawutils
 from caselawclient.Client import api_client
 from caselawclient.client_helpers.search_helpers import (
     search_judgments_and_parse_response,
@@ -8,7 +9,10 @@ from caselawclient.errors import DocumentNotFoundError
 from caselawclient.models.documents import Document
 from caselawclient.search_parameters import SearchParameters
 from django.http import Http404
+from django.views.generic import TemplateView
 
+from judgments.utils import editors_dict, set_document_type_and_link
+from judgments.utils.link_generators import build_jira_create_link
 from judgments.utils.paginator import paginator
 
 ALLOWED_ORDERS = ["date", "-date"]
@@ -52,4 +56,32 @@ def get_document_by_uri_or_404(uri: str) -> Document:
     try:
         return api_client.get_document_by_uri(uri)
     except DocumentNotFoundError:
-        raise Http404(f"Document not found at {uri}")
+        raise Http404(f"Judgment not found at {uri}")
+
+
+class DocumentView(TemplateView):
+    def get_context_data(self, **kwargs):
+        document_uri = kwargs["document_uri"]
+
+        document = get_document_by_uri_or_404(document_uri)
+
+        context = super().get_context_data(**kwargs)
+        context["document_uri"] = document_uri
+
+        context["document"] = document
+
+        # TODO: Remove this once we fully deprecate 'judgment' contexts
+        context["judgment"] = document
+
+        context["page_title"] = document.name
+        context["courts"] = caselawutils.courts.get_all()
+
+        context.update({"editors": editors_dict()})
+
+        context["jira_create_link"] = build_jira_create_link(
+            document=document, request=self.request
+        )
+
+        context = set_document_type_and_link(context, document_uri)
+
+        return context
